[libc++] Merge is_member{,_object,_function}_pointer.h
